added push pull and add remtoe, cleaned up ui, added status updates to the bottom bar

+++ src/js/mutations/MoveLabbookFileMutation.js
@@ -26,17 +26,46 @@ const mutation = graphql`
 
 let tempID = 0;
 
-function sharedUpdater(store, labbookID, deletedID, connectionKey) {
+function sharedDeleteUpdater(store, labbookID, deletedID, connectionKey) {
   const userProxy = store.get(labbookID);
   const conn = RelayRuntime.ConnectionHandler.getConnection(
     userProxy,
     connectionKey,
   );
 
-  RelayRuntime.ConnectionHandler.deleteNode(
-    conn,
-    deletedID
-  );
+  if(conn){
+    RelayRuntime.ConnectionHandler.deleteNode(
+      conn,
+      deletedID
+    );
+  }
+}
+
+function sharedUpdater(store, labbookId, connectionKey, node) {
+  const labbookProxy = store.get(labbookId);
+
+
+  if(labbookProxy){
+    const conn = RelayRuntime.ConnectionHandler.getConnection(
+      labbookProxy,
+      connectionKey
+    );
+
+
+    if(conn){
+      const newEdge = RelayRuntime.ConnectionHandler.createEdge(
+        store,
+        conn,
+        node,
+        "newLabbookFileEdge"
+      )
+
+      RelayRuntime.ConnectionHandler.insertEdgeAfter(
+        conn,
+        newEdge
+      );
+    }
+  }
 }
 
 
@@ -51,8 +80,7 @@ export default function MoveLabbookFileMutation(
   section,
   callback
 ) {
-  console.log(srcPath,
-  dstPath)
+
   const variables = {
     input: {
       owner,
@@ -88,7 +116,7 @@ export default function MoveLabbookFileMutation(
       onError: err => console.error(err),
       optimisticUpdater: (store) => {
 
-        sharedUpdater(store, labbookId, edge.node.id, connectionKey);
+        // sharedUpdater(store, labbookId, edge.node.id, connectionKey);
         const id = 'client:newFileMove:'+ tempID++;
         const userProxy = store.get(labbookId);
 
@@ -121,8 +149,20 @@ export default function MoveLabbookFileMutation(
           );
         }
       },
-      updater: (store) => {
-        sharedUpdater(store, labbookId, edge.node.id, connectionKey);
+      updater: (store, response) => {
+        console.log(srcPath)
+        sharedDeleteUpdater(store, labbookId, edge.node.id, connectionKey);
+
+        // if(response.moveLabbookFile && response.moveLabbookFile.newLabbookFileEdge){
+        //   const node = store.create('client:newFile:'+ (tempID++), 'CodeFile')
+        //   node.setValue(response.moveLabbookFile.newLabbookFileEdge.node.id, "id")
+        //   node.setValue(false, 'isDir')
+        //   node.setValue(response.moveLabbookFile.newLabbookFileEdge.node.key, 'key')
+        //   node.setValue(response.moveLabbookFile.newLabbookFileEdge.node.modifiedAt, 'modifiedAt')
+        //   node.setValue(response.moveLabbookFile.newLabbookFileEdge.node.size, 'size')
+        //
+        //   sharedUpdater(store, labbookId, connectionKey, node)
+        // }
       }
 
     },
